test: 테스트 코드 클래스 통합 및 데이터 타입 수정
